+++ src/de/uvwxy/footpath2/map/IndoorLocation.java
@@ -244,12 +244,12 @@ public class IndoorLocation extends Location implements DrawToCanvas {
 	private int[] apix = { 0, 0 };
 
 	@Override
-	public void drawToCanvas(Canvas canvas, IndoorLocation center, Rect boundingBox, double pixelsPerMeterOrMaxValue,
+	public void drawToCanvas(Canvas canvas, IndoorLocation center, Rect boundingBox, float pixelsPerMeterOrMaxValue,
 			Paint pLine, Paint pDots) {
 		w = boundingBox.width() / 2 + boundingBox.left;
 		h = boundingBox.height() / 2 + boundingBox.top;
 
 		apix = GeoUtils.convertToPixelLocation(this, center, pixelsPerMeterOrMaxValue);
-		canvas.drawCircle(w + apix[0], h + apix[1], (float) (pixelsPerMeterOrMaxValue * 0.5f), pDots);
+		canvas.drawCircle(w + apix[0], h + apix[1], pixelsPerMeterOrMaxValue * 0.5f, pDots);
 	}
 }
